Deleted original backend, renamed chronochords-user-service to backend.
