+++ backend/.gitignore
@@ -1,4 +1,5 @@
 HELP.md
+application.properties
 target/
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
