--- backend/src/main/java/com/ChronoChords/ChronoChords/ChronoChordsApplication.java
@@ -1,13 +1,0 @@
-package com.ChronoChords.ChronoChords;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ChronoChordsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ChronoChordsApplication.class, args);
-	}
-
-}
