--- backend/src/main/java/com/ChronoChords/ChronoChords/config/WebConfig.java
@@ -1,19 +1,0 @@
-package com.ChronoChords.ChronoChords.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/api/registration/**")
-                .allowedOrigins("http://localhost:3000")  // or "*" to allow all, but be cautious with this in production
-                .allowedMethods("GET", "POST", "PUT", "DELETE")
-                .allowedHeaders("*")
-                .allowCredentials(true)
-                .maxAge(3600); // Max age for the preflight request caching in seconds
-    }
-}
