--- backend/src/main/java/com/ChronoChords/ChronoChords/model/User.java
@@ -1,55 +1,0 @@
-package com.ChronoChords.ChronoChords.model;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "users")
-public class User {
-
-    @Id
-    @Column(name="user_id", length = 45)
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
-	
-	@Column(name = "email")
-	private String email;
-
-	@Column(name = "password")
-	private String password;
-	
-	public User() {
-    }
-    public User(long id, String email, String password) {
-        super();
-        this.id = id;
-        this.password = password;
-        this.email = email;
-    }
-
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getPassword(){
-        return password;
-    }
-
-    public void setPassword(String password){
-        this.password = password;
-    }
-
-    public String getEmail(){
-        return email;
-    }
-    
-    public void setEmail(String email){
-        this.email = email;
-    }
-}
-
-
