--- backend/src/main/java/com/ChronoChords/ChronoChords/model/loginResponse.java
@@ -1,23 +1,0 @@
-package com.ChronoChords.ChronoChords.model;
-
-public class loginResponse {
-    String message;
-    Boolean status;
-    public String getMessage() {
-        return message;
-    }
-    public void setMessage(String message) {
-        this.message = message;
-    }
-    public Boolean getStatus() {
-        return status;
-    }
-    public void setStatus(Boolean status) {
-        this.status = status;
-    }
-    public loginResponse(String message, Boolean status) {
-        this.message = message;
-        this.status = status;
-    }
-}
-
