+++ backend/src/main/java/com/chronochords/backend/DTO/UserDTO.java
@@ -1,0 +1,56 @@
+package com.chronochords.backend.DTO;
+import jakarta.validation.constraints.*;
+
+public class UserDTO {
+    private int userid;
+    @NotBlank(message = "Username is required")
+    @Size(min = 3, message = "Username must be at least 3 characters long")
+    private String username;
+    @NotBlank(message = "Email is required")
+    @Email(message = "Email should be valid")
+    private String email;
+    @NotBlank(message = "Password is required")
+    @Size(min = 4, message = "Password must be at least 4 characters long")
+    private String password;
+    public UserDTO() {
+    }
+    public UserDTO(int userid, String username, String email, String password) {
+        this.userid = userid;
+        this.username = username;
+        this.email = email;
+        this.password = password;
+    }
+
+    public int getUserid() {
+        return userid;
+    }
+
+    public void setUserid(int userid) {
+        this.userid = userid;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
+
