+++ backend/src/main/java/com/chronochords/backend/Repository/UserRepo.java
@@ -1,0 +1,18 @@
+package com.chronochords.backend.Repository;
+
+import com.chronochords.backend.Entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.stereotype.Repository;
+import java.util.Optional;
+@EnableJpaRepositories
+@Repository
+public interface UserRepo extends JpaRepository<User, Long>
+{
+    Optional<User> findOneByEmailAndPassword(String email, String password);
+    User findByEmail(String email);
+
+    Boolean existsByEmail(String email);
+
+    Boolean existsByUsername(String username);
+}
