+++ backend/src/main/java/com/chronochords/backend/Service/UserService.java
@@ -1,0 +1,9 @@
+package com.chronochords.backend.Service;
+
+import com.chronochords.backend.DTO.UserDTO;
+import com.chronochords.backend.DTO.LoginDTO;
+import com.chronochords.backend.payload.response.LoginMessage;
+public interface UserService {
+    String addUser(UserDTO userDTO);
+    LoginMessage loginUser(LoginDTO loginDTO);
+}
