+++ backend/src/main/resources/application.properties
@@ -1,8 +1,14 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/dbkms?createDatabaseIfNotExist=true
+spring.application.name=ChronoChords
+server.port=8085
 
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+spring.datasource.url=jdbc:mysql://localhost:3306/chronochords?createDatabaseIfNotExist=true
 spring.datasource.username=root
-spring.datasource.password=rootroot
+spring.datasource.password=dreamcoreOBLXKQ!9501
 
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQLDialect
-spring.jpa.hibernate.ddl-auto = update
+#jpa vendor adapter configuration
+spring.jpa.database-platform=org.hibernate.dialect.MySQL57Dialect
+spring.jpa.generate-ddl=true
 spring.jpa.show-sql=true
