--- backend/src/test/java/com/ChronoChords/ChronoChords/ChronoChordsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ChronoChords.ChronoChords;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ChronoChordsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
