--- chronochords-user-service/src/main/java/com/chronochords/chronochordsuserservice/ChronochordsUserServiceApplication.java
@@ -1,14 +1,0 @@
-package com.chronochords.chronochordsuserservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-
-@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
-public class ChronochordsUserServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ChronochordsUserServiceApplication.class, args);
-	}
-
-}
