--- chronochords-user-service/src/main/java/com/chronochords/chronochordsuserservice/DTO/LoginDTO.java
@@ -1,30 +1,0 @@
-package com.chronochords.chronochordsuserservice.DTO;
-
-public class LoginDTO {
-    private String email;
-    private String password;
-
-    public LoginDTO() {
-    }
-
-    public LoginDTO(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
