--- chronochords-user-service/src/main/java/com/chronochords/chronochordsuserservice/Repository/UserRepo.java
@@ -1,18 +1,0 @@
-package com.chronochords.chronochordsuserservice.Repository;
-
-import com.chronochords.chronochordsuserservice.Entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.stereotype.Repository;
-import java.util.Optional;
-@EnableJpaRepositories
-@Repository
-public interface UserRepo extends JpaRepository<User, Long>
-{
-    Optional<User> findOneByEmailAndPassword(String email, String password);
-    User findByEmail(String email);
-
-    Boolean existsByEmail(String email);
-
-    Boolean existsByUsername(String username);
-}
