--- chronochords-user-service/src/test/java/com/chronochords/chronochordsuserservice/ChronochordsUserServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.chronochords.chronochordsuserservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ChronochordsUserServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
