+++ frontend/src/pages/MainGame.js
@@ -1,7 +1,7 @@
 import '../App.css';
 import React, { useState } from 'react';
 import Slider from '../components/Slider';
-import MusicPlayer from '../components/Musicplayer';
+import MusicPlayer from '../components/MusicPlayer';
 import RoundCount from '../components/RoundCount';
 import Popup from '../components/Popup';
 import ScoreDisplay from '../components/ScoreDisplay';
